Extract brace handlers from pattern builder
